requirements/dev.txt
# Development dependencies
-r base.txt

# Testing
pytest>=7.4.0
pytest-django>=4.5.2
pytest-cov>=4.1.0
pytest-xdist>=3.3.1
factory-boy>=3.3.0

# Code quality
black>=23.10.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.6.0
django-stubs>=4.2.6
djangorestframework-stubs>=3.14.4

# Pre-commit hooks
pre-commit>=3.5.0

# Security scanning
bandit>=1.7.6
detect-secrets>=1.4.0

# Development tools
ipython>=8.16.0
ipdb>=0.13.13
django-extensions>=3.2.3
